requirements.txt
numpy
pandas
matplotlib
sklearn
seaborn
scipy
statsmodels
nltk
TextBlob
spacy
scikit-learn
TA-Lib
pynance
plotly
mplfinance
jupyterlab
yfinance
requests
gdown
gensim
python-dateutil
vaderSentiment